Cleaned up code.
